+++ src/lib.rs
@@ -21,9 +21,33 @@
 pub use crate::server::AstarteMessageHub;
 
 pub mod astarte;
-pub(crate) mod cache;
+pub mod cache;
 pub mod config;
-mod device;
 pub mod error;
+#[cfg(feature = "security-events")]
+pub mod events;
 mod messages;
 mod server;
+pub mod store;
+
+#[cfg(test)]
+pub(crate) mod tests {
+    use insta::assert_snapshot;
+
+    macro_rules! with_settings {
+        ($asserts:block) => {
+            ::insta::with_settings!({
+                snapshot_path => concat!(env!("CARGO_MANIFEST_DIR"), "/snapshots")
+            }, $asserts);
+        };
+    }
+
+    pub(crate) use with_settings;
+
+    #[test]
+    fn use_macro() {
+        self::with_settings!({
+            assert_snapshot!("using the macro");
+        });
+    }
+}